+++ tests/test_form.py
@@ -136,25 +136,23 @@ def test_axi():
 
     for parallel in [False, True]:
 
-        for jit in [False, True]:
+        L = fe.IntegralForm(P, u, r.dV)
+        x = L.integrate(parallel=parallel)
 
-            L = fe.IntegralForm(P, u, r.dV)
-            x = L.integrate(parallel=parallel, jit=jit)
+        b = L.assemble(x, parallel=parallel).toarray()
+        assert b.shape == (r.mesh.ndof, 1)
 
-            b = L.assemble(x, parallel=parallel).toarray()
-            assert b.shape == (r.mesh.ndof, 1)
+        b = L.assemble(parallel=parallel).toarray()
+        assert b.shape == (r.mesh.ndof, 1)
 
-            b = L.assemble(parallel=parallel).toarray()
-            assert b.shape == (r.mesh.ndof, 1)
+        a = fe.IntegralForm(A, u, r.dV, u, grad_v=[True], grad_u=[True])
+        y = a.integrate(parallel=parallel)
 
-            a = fe.IntegralForm(A, u, r.dV, u, grad_v=[True], grad_u=[True])
-            y = a.integrate(parallel=parallel, jit=jit)
+        K = a.assemble(y, parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.ndof)
 
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.ndof)
-
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.ndof)
+        K = a.assemble(parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.ndof)
 
 
 def test_linearform():
@@ -163,21 +161,19 @@ def test_linearform():
 
     for parallel in [False, True]:
 
-        for jit in [False, True]:
+        L = fe.IntegralForm(P, u, r.dV, grad_v=[True])
+        x = L.integrate(parallel=parallel)
+        b = L.assemble(x, parallel=parallel).toarray()
+        assert b.shape == (r.mesh.ndof, 1)
+        b = L.assemble(parallel=parallel).toarray()
+        assert b.shape == (r.mesh.ndof, 1)
 
-            L = fe.IntegralForm(P, u, r.dV, grad_v=[True])
-            x = L.integrate(parallel=parallel, jit=jit)
-            b = L.assemble(x, parallel=parallel).toarray()
-            assert b.shape == (r.mesh.ndof, 1)
-            b = L.assemble(parallel=parallel, jit=jit).toarray()
-            assert b.shape == (r.mesh.ndof, 1)
-
-            L = fe.IntegralForm(p.extract(grad=False), p, r.dV, grad_v=[False])
-            x = L.integrate(parallel=parallel, jit=jit)
-            b = L.assemble(x, parallel=parallel, jit=jit).toarray()
-            assert b.shape == (r.mesh.npoints, 1)
-            b = L.assemble(parallel=parallel, jit=jit).toarray()
-            assert b.shape == (r.mesh.npoints, 1)
+        L = fe.IntegralForm(p.extract(grad=False), p, r.dV, grad_v=[False])
+        x = L.integrate(parallel=parallel)
+        b = L.assemble(x, parallel=parallel).toarray()
+        assert b.shape == (r.mesh.npoints, 1)
+        b = L.assemble(parallel=parallel).toarray()
+        assert b.shape == (r.mesh.npoints, 1)
 
 
 def test_linearform_broadcast():
@@ -186,21 +182,19 @@ def test_linearform_broadcast():
 
     for parallel in [False, True]:
 
-        for jit in [False, True]:
+        L = fe.IntegralForm(P, u, r.dV, grad_v=[True])
+        x = L.integrate(parallel=parallel)
+        b = L.assemble(x, parallel=parallel).toarray()
+        assert b.shape == (r.mesh.ndof, 1)
+        b = L.assemble(parallel=parallel).toarray()
+        assert b.shape == (r.mesh.ndof, 1)
 
-            L = fe.IntegralForm(P, u, r.dV, grad_v=[True])
-            x = L.integrate(parallel=parallel, jit=jit)
-            b = L.assemble(x, parallel=parallel, jit=jit).toarray()
-            assert b.shape == (r.mesh.ndof, 1)
-            b = L.assemble(parallel=parallel, jit=jit).toarray()
-            assert b.shape == (r.mesh.ndof, 1)
-
-            L = fe.IntegralForm(p.extract(grad=False), p, r.dV, grad_v=[False])
-            x = L.integrate(parallel=parallel, jit=jit)
-            b = L.assemble(x, parallel=parallel, jit=jit).toarray()
-            assert b.shape == (r.mesh.npoints, 1)
-            b = L.assemble(parallel=parallel, jit=jit).toarray()
-            assert b.shape == (r.mesh.npoints, 1)
+        L = fe.IntegralForm(p.extract(grad=False), p, r.dV, grad_v=[False])
+        x = L.integrate(parallel=parallel)
+        b = L.assemble(x, parallel=parallel).toarray()
+        assert b.shape == (r.mesh.npoints, 1)
+        b = L.assemble(parallel=parallel).toarray()
+        assert b.shape == (r.mesh.npoints, 1)
 
 
 def test_bilinearform():
@@ -209,21 +203,19 @@ def test_bilinearform():
 
     for parallel in [False, True]:
 
-        for jit in [False, True]:
+        a = fe.IntegralForm(A, u, r.dV, u)
+        y = a.integrate(parallel=parallel)
+        K = a.assemble(y, parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.ndof)
+        K = a.assemble(parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.ndof)
 
-            a = fe.IntegralForm(A, u, r.dV, u)
-            y = a.integrate(parallel=parallel, jit=jit)
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.ndof)
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.ndof)
-
-            a = fe.IntegralForm(P, u, r.dV, p, [True], [False])
-            y = a.integrate(parallel=parallel, jit=jit)
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.npoints)
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.npoints)
+        a = fe.IntegralForm(P, u, r.dV, p, [True], [False])
+        y = a.integrate(parallel=parallel)
+        K = a.assemble(y, parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.npoints)
+        K = a.assemble(parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.npoints)
 
 
 def test_bilinearform_broadcast():
@@ -232,30 +224,28 @@ def test_bilinearform_broadcast():
 
     for parallel in [False, True]:
 
-        for jit in [False, True]:
+        a = fe.IntegralForm(A, u, r.dV, u, [True], [True])
+        y = a.integrate(parallel=parallel)
+        K = a.assemble(y, parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.ndof)
+        K = a.assemble(parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.ndof)
 
-            a = fe.IntegralForm(A, u, r.dV, u, [True], [True])
-            y = a.integrate(parallel=parallel, jit=jit)
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.ndof)
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.ndof)
+        a = fe.IntegralForm(P, u, r.dV, p, [True], [False])
+        y = a.integrate(parallel=parallel)
+        K = a.assemble(y, parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.npoints)
+        K = a.assemble(parallel=parallel).toarray()
+        assert K.shape == (r.mesh.ndof, r.mesh.npoints)
 
-            a = fe.IntegralForm(P, u, r.dV, p, [True], [False])
-            y = a.integrate(parallel=parallel, jit=jit)
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.npoints)
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.ndof, r.mesh.npoints)
-
-            q = p.extract(grad=False)
-            f = fe.math.dya(q, q, mode=1)
-            a = fe.IntegralForm(f, p, r.dV, p, [False], [False])
-            y = a.integrate(parallel=parallel, jit=jit)
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.npoints, r.mesh.npoints)
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
-            assert K.shape == (r.mesh.npoints, r.mesh.npoints)
+        q = p.extract(grad=False)
+        f = fe.math.dya(q, q, mode=1)
+        a = fe.IntegralForm(f, p, r.dV, p, [False], [False])
+        y = a.integrate(parallel=parallel)
+        K = a.assemble(y, parallel=parallel).toarray()
+        assert K.shape == (r.mesh.npoints, r.mesh.npoints)
+        K = a.assemble(parallel=parallel).toarray()
+        assert K.shape == (r.mesh.npoints, r.mesh.npoints)
 
 
 def test_mixed():
@@ -264,43 +254,39 @@ def test_mixed():
 
     for parallel in [False, True]:
 
-        for jit in [False, True]:
+        a = fe.IntegralForm(A, v, r.dV, v)
+        y = a.integrate(parallel=parallel)
+        K = a.assemble(y, parallel=parallel).toarray()
+        K = a.assemble(parallel=parallel).toarray()
 
-            a = fe.IntegralForm(A, v, r.dV, v)
-            y = a.integrate(parallel=parallel, jit=jit)
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
+        z = r.mesh.ndof + 2 * r.mesh.npoints
+        assert K.shape == (z, z)
 
-            z = r.mesh.ndof + 2 * r.mesh.npoints
-            assert K.shape == (z, z)
+        L = fe.IntegralForm(f, v, r.dV)
+        x = L.integrate(parallel=parallel)
+        b = L.assemble(x, parallel=parallel).toarray()
+        b = L.assemble(parallel=parallel).toarray()
 
-            L = fe.IntegralForm(f, v, r.dV)
-            x = L.integrate(parallel=parallel, jit=jit)
-            b = L.assemble(x, parallel=parallel, jit=jit).toarray()
-            b = L.assemble(parallel=parallel, jit=jit).toarray()
-
-            assert b.shape == (z, 1)
+        assert b.shape == (z, 1)
 
     r, v, f, A = pre_axi_mixed()
 
     for parallel in [False, True]:
 
-        for jit in [False, True]:
+        a = fe.IntegralForm(A, v, r.dV, v)
+        y = a.integrate(parallel=parallel)
+        K = a.assemble(y, parallel=parallel).toarray()
+        K = a.assemble(parallel=parallel).toarray()
 
-            a = fe.IntegralForm(A, v, r.dV, v)
-            y = a.integrate(parallel=parallel, jit=jit)
-            K = a.assemble(y, parallel=parallel, jit=jit).toarray()
-            K = a.assemble(parallel=parallel, jit=jit).toarray()
+        z = r.mesh.ndof + 2 * r.mesh.npoints
+        assert K.shape == (z, z)
 
-            z = r.mesh.ndof + 2 * r.mesh.npoints
-            assert K.shape == (z, z)
+        L = fe.IntegralForm(f, v, r.dV)
+        x = L.integrate(parallel=parallel)
+        b = L.assemble(x, parallel=parallel).toarray()
+        b = L.assemble(parallel=parallel).toarray()
 
-            L = fe.IntegralForm(f, v, r.dV)
-            x = L.integrate(parallel=parallel, jit=jit)
-            b = L.assemble(x, parallel=parallel, jit=jit).toarray()
-            b = L.assemble(parallel=parallel, jit=jit).toarray()
-
-            assert b.shape == (z, 1)
+        assert b.shape == (z, 1)
 
 
 if __name__ == "__main__":