+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file. The format 
 
 ## [Unreleased]
 
+### Changed
+- Add `einsumt` to requirements.
+
+### Removed
+- Remove `jit`-compilation of forms (`parallel` is the preferred method).
+- Remove unused `tools.check()`.
+
 ## [6.4.0] - 2023-04-01
 
 ### Added